requirements.txt
asgiref==3.5.2
astroid==2.12.9
autopep8==1.7.0
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.1.1
django-filter==22.1
django-on-heroku==1.1.2
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pylint==2.15.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
whitenoise==6.2.0
wrapt==1.14.1
